Feat/#120/icon market

+++ src/main/java/com/timeToast/timeToast/dto/icon/icon_group/request/IconGroupPostRequest.java
@@ -7,7 +7,7 @@ import com.timeToast.timeToast.domain.icon.icon_group.IconGroup;
 public record IconGroupPostRequest (
 
         String name,
-        long price,
+        int price,
         IconType iconType,
         IconBuiltin iconBuiltin
